chore: change home page, use open source instead
