+++ assets/@theme/scss/_dark.scss
@@ -1,3 +1,5 @@
+
+/// dark
 $dark_primary_color: #212121;
 $dark_secondary_color: #BABABA;
 $dark_white_color: #fafafa;
@@ -13,15 +15,15 @@ $white: #fff;
     ul.left {
       color: $dark_white_color;
       a {
-        font-weight: 100;
+        font-weight: 300;
         transition: all .26s linear;
 
         &:hover {
-          font-weight: 300;
+          font-weight: 800;
         }
       }
       a.active {
-        font-weight: 400;
+        font-weight: 800;
       }
     }
 
@@ -88,6 +90,33 @@ $white: #fff;
     .icon-2x::before {
       font-size: 2em;
       color: #fff;
+    }
+  }
+
+  .contact {
+    form {
+      p {
+        color: $dark_secondary_color;
+      }
+
+      input::placeholder, textarea::placeholder {
+        color:$dark_secondary_color;
+      }
+
+      input, textarea {
+        background-color:rgba(21, 21, 21, .82);
+        color:$white;
+      }
+    }
+
+    button {
+      background: rgba(21, 21, 21, .82);
+      color: $white;
+      padding: 12px 30px;
+      font-family: 'Gothic A1', sans-serif;
+      &:hover {
+        background: rgba(15, 15, 15, 1);
+      }
     }
   }
