+++ assets/@theme/scss/_light.scss
@@ -1,3 +1,5 @@
+
+/// light
 $light_primary_color: #f0f0f0;
 $light_secondary_color: #212121;
 $light_white_color: #fafafa;
@@ -11,14 +13,14 @@ $light_white_color: #fafafa;
   nav {
     ul.left {
       a {
-        font-weight: 100;
+        font-weight: 300;
 
         &:hover {
-          font-weight: 300;
+          font-weight: 800;
         }
       }
       a.active {
-        font-weight: 400;
+        font-weight: 800;
       }
     }
 
@@ -118,6 +120,28 @@ $light_white_color: #fafafa;
       }
     }
 
+  }
+
+  .contact {
+    form {
+      p {
+        color: $light_secondary_color;
+      }
+
+      input::placeholder, textarea::placeholder {
+        color: #444;
+      }
+    }
+
+    button {
+      background: $light_secondary_color;
+      color: $light_primary_color;
+      padding: 12px 30px;
+      font-family: 'Gothic A1', sans-serif;
+      &:hover {
+        background:rgba(21, 21, 21, .82);
+      }
+    }
   }
 
   .app-title.sub-title {
