+++ assets/@theme/scss/index.scss
@@ -48,3 +48,40 @@ html {
   -webkit-font-smoothing: antialiased;
   box-sizing: border-box;
 }
+
+
+nav {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+
+  ul.left {
+    span.separator {
+      margin: 0 5px;
+    }
+  }
+  ul.left, ul.right {
+    a {
+      outline: none;
+      display: flex;
+    }
+  }
+}
+.gothic-font {
+  font-family: "Gothic A1", 'Source Sans Pro', -apple-system, BlinkMacSystemFont, 'Segoe UI',
+    Roboto, 'Helvetica Neue', Arial, sans-serif;
+}
+.mobile-menu {
+  backdrop-filter: blur(2px);
+  button.absolute {
+    right: 10px;
+    top: 10px;
+    width: 35px;
+    display: flex;
+    align-items: center;
+    justify-content:center;
+    height: 35px;
+    border-radius:25px;
+    align-content: center;
+  }
+}
