+++ pages/index/index.vue
@@ -1,107 +1,82 @@
 <template>
-  <div class="articles">
-    <app-title>Blog Articles  😊</app-title>
-    <section class="grid grid-cols-1 sm:grid-cols-2 gap-5 mx-auto md:overflow-y-auto overflow-x-hidden md:h-90 md:pt-6">
-      <article
-        v-for="article in articles"
-        :key="article.link"
-        :style="{ backgroundImage: `url(${article.image})` }"
-        class="relative"
-      >
-        <nuxt-link class="py-6 px-6" :to="article.link">
-          <header>
-            <span class="bottom-0 mb-6">{{ article.date | formateDate }}</span>
-          </header>
-          <h2 class="text-2xl">
-            {{ article.title.trim() | capitalize }}
-          </h2>
-          <!-- <p class="my-3">{{ article.description.trim().slice(0, 100) }} ...</p> -->
-          <div class="categories flex-wrap flex">
-            <p class="mr-2 text-xs rounded capitalize">
-              {{ article.categories }}
-            </p>
-          </div>
-        </nuxt-link>
-      </article>
+  <div
+    v-infinite-scroll="loadMore"
+    infinite-scroll-disabled="busy"
+    infinite-scroll-distance="100"
+    class="projects"
+  >
+    <app-title>
+      Open source projects 🤜🏽 <br>
+      <small class="font-light text-gray-200 gothic-font"> Find one and contribute 😄</small>
+    </app-title>
+    <section
+      v-infinite-scroll="loadMore"
+      infinite-scroll-disabled="busy"
+      infinite-scroll-distance="100"
+      class="grid grid-cols-1 sm:grid-cols-2 gap-5 px-3 mx-auto md:overflow-y-auto overflow-x-hidden md:h-90 md:pt-6"
+    >
+      <app-project-item
+        v-for="repo of repos"
+        :key="repo.id"
+        :repo="repo"
+      />
     </section>
   </div>
 </template>
 
 <script>
-import { formateDate, capitalize } from '~/filters'
+const loadRepositories = async (page = 1) => {
+  const repos = await fetch('https://api.github.com/users/acidiney/repos?sort=created&page=' + page)
+    .then(req => req.json())
+    .then(repo => repo.filter(r => !r.fork))
+    .then(repo => repo.sort((a, b) => a.stargazers_count > b.stargazers_count ? -1 : 1))
+
+  return repos
+}
 
 export default {
-  name: 'Articles',
-  filters: {
-    capitalize,
-    formateDate
-  },
-  async asyncData ({ $content }) {
-    const articles = await $content()
-      .only(['title', 'slug', 'date', 'image', 'categories', 'link'])
-      .sortBy('date', 'desc')
-      .limit(4)
-      .fetch()
-
+  name: 'Projects',
+  async asyncData (context) {
+    const repos = await loadRepositories()
     return {
-      articles
+      repos
+    }
+  },
+  data () {
+    return {
+      moreRepos: [],
+      page: 1
+    }
+  },
+  computed: {
+    listRepos () {
+      return [].concat(...this.repos).concat(...this.moreRepos)
+    }
+  },
+  methods: {
+    async loadMore () {
+      this.page++
+      await loadRepositories(this.page)
+        .then((repos) => {
+          this.moreRepos.push(repos)
+        })
     }
   }
 }
 </script>
 
 <style lang="scss" scoped>
-.articles article {
-  height: 280px;
-  background-size: cover;
-  margin-bottom: 20px;
-  border-radius: 25px;
-  transition: all .26s linear;
-  cursor: pointer;
-  overflow: hidden;
-}
 
 .md\:h-90 {
-    @media (min-width:768px) {
-      max-height: 86vh;
-      -ms-overflow-style: none;  /* IE and Edge */
-      scrollbar-width: none;  /* Firefox */
+  @media (min-width:768px) {
+    height: 86vh;
+    padding-bottom: 5em;
+    -ms-overflow-style: none;  /* IE and Edge */
+    scrollbar-width: none;  /* Firefox */
 
     &::webkit-scrollbar {
       display: none;
     }
   }
 }
-
-.articles article a {
-  display: block;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-end;
-}
-
-.articles article a header .categories p {
-  font-weight: 100;
-  font-family: 'Gothic A1';
-  /* color: white; */
-  text-transform: uppercase;
-}
-
-.articles article a h2 {
-  font-weight: 400;
-  /* color: #fff; */
-  font-family: 'Roboto', sans-serif;
-}
-
-.articles article a p {
-  color:rgba(148, 148, 149, 0.9);
-}
-
-.articles article a span {
-  /* color: #fff; */
-  font-family: 'Gothic A1', sans-serif;
-  font-weight: 100;
-}
-
 </style>
