+++ plugins/project-item.component.js
@@ -1,0 +1,4 @@
+import Vue from 'vue'
+import ProjectItem from '~/components/project-item'
+
+Vue.component('app-project-item', ProjectItem)
